added segmented config wrapper
