+++ Runtime/ConfigurationInstaller.cs
@@ -36,6 +36,7 @@
             services.Add<IConfigApplyCheckerFactory>().ImplementedBy<ConfigApplyCheckerFactory>().AsSingleton();
             services.Add<IConfigMetaProvider>().ImplementedByInstance(configMetaProvider);
             services.Add<IConfigsRegistry>().ImplementedBy<ConfigsRegistry>().AsSingleton();
+            services.Add<IConfigValueResolver>().ImplementedBy<SegmentedConfigsResolver>().AsSingleton();
             services.Add<IConfigsFactory>().ImplementedBy<ConfigsFactory>().AsSingleton();
             services.Add<IConfigValuesInjector, IConfigMetadataExtractor>().ImplementedBy<MultiFormatConfigValuesHandler>().AsSingleton();
             services.Add<IAppliedConfigsHolder>().ImplementedBy<AppliedConfigsHolder>().AsSingleton();
