+++ Runtime/General/Fetch/CombinedConfigsFetcher.cs
@@ -7,8 +7,6 @@
     using Core.Features;
     using Core.Version;
     using Cysharp.Threading.Tasks;
-    using Newtonsoft.Json;
-    using UnityEngine;
     using Zenject;
     using Logger = Logs.Logger;
 
@@ -41,7 +39,8 @@
         [Inject]
         public IConfigMetadataExtractor MetadataExtractor { get; set; }
 
-        private readonly IDictionary<string, string> values = new Dictionary<string, string>();
+        [Inject]
+        public IConfigValueResolver ConfigsResolver { get; set; }
 
         private IConfigsFetcher remoteFetcher;
         
@@ -121,7 +120,7 @@
                 if (applyChecker.ShouldApply())
                 {
                     Logger.Log("Config", $"Config '{key}' is applied from remote value{logSuffix}{remoteValue}");
-                    values[key] = remoteValue;
+                    PutValue(key, remoteValue);
 
                     if (metadata.HasActivationEvent)
                     {
@@ -157,7 +156,7 @@
                 if (appliedConfig != null)
                 {
                     Logger.Log("Config", $"Config '{key}' is taken from cache{logSuffix}{appliedConfig}");
-                    values[key] = appliedConfig;
+                    PutValue(key, appliedConfig);
                     return;
                 }
 
@@ -171,24 +170,25 @@
                 Logger.Log("Config", $"Config '{key}' is taken from local value{logSuffix}{appliedConfig}");
                 if (appliedConfig != null)
                 {
-                    values[key] = appliedConfig;
+                    PutValue(key, appliedConfig);
                 }
             }
         }
 
+        private void PutValue(string key, string value)
+        {
+            ConfigsResolver.PutValue(key, value);
+        }
+
         public IEnumerable<string> GetKeys()
         {
-            return values.Keys;
+            return ConfigsResolver.GetKeys();
         }
 
         public string GetValue(string key)
         {
-            return values[key];
+            return ConfigsResolver.GetValue(key);
         }
 
-        public bool HasKey(string key)
-        {
-            return values.ContainsKey(key);
-        }
     }
 }
