+++ Runtime/General/Fetch/SegmentedConfig.cs
@@ -1,0 +1,17 @@
+namespace Unibrics.Configuration.General
+{
+    public class SegmentedConfig
+    {
+        public string DefaultValue { get; private set; }
+        
+        public void AddDefaultValue(string value)
+        {
+            DefaultValue = value;
+        }
+
+        public void AddSegmentValue(string segment, string value)
+        {
+            
+        }
+    }
+}
