+++ Runtime/General/Fetch/SegmentedConfig.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 8554a44ee19049f0b512bcd8aed288b5
+timeCreated: 1719823535
