+++ Runtime/General/Fetch/SegmentedConfigsResolver.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: a97a4a745b9e43d28f206d21e31521a3
+timeCreated: 1719824099
